create service details event and flow
