+++ app/src/main/java/com/example/e_event/view/main/MainActivityViewModel.kt
@@ -13,7 +13,6 @@ import retrofit2.Response
 class MainActivityViewModel() : ViewModel() {
 
     var obj: MutableLiveData<List<Event>> = MutableLiveData<List<Event>>()
-    var eventId: MutableLiveData<Event> = MutableLiveData<Event>()
     var msg: MutableLiveData<String> = MutableLiveData<String>()
     val service: EventAPI = Network.getInstance(EventAPI::class.java).build("http://5f5a8f24d44d640016169133.mockapi.io/api/")
 
@@ -31,20 +30,4 @@ class MainActivityViewModel() : ViewModel() {
 
         })
     }
-
-    fun checkDetails(id: Int) {
-        service.getDetailsEvent(id).enqueue(object : Callback<Event> {
-            override fun onResponse(call: Call<Event>?, response: Response<Event>?) {
-                if (response != null) {
-                    eventId.value = response.body()
-                }
-            }
-
-            override fun onFailure(call: Call<Event>?, t: Throwable?) {
-                Log.w("Error", "CAIU")
-            }
-
-        })
-    }
-
 }
