+++ app/src/main/res/layout/activity_detail.xml
@@ -25,6 +25,7 @@
         <TextView
             android:id="@+id/tvLatitude"
             tools:text="-30.0392981"
+            android:textColor="@color/black"
             android:padding="2dp"
             android:elevation="20dp"
             android:layout_width="wrap_content"
@@ -37,6 +38,7 @@
         <TextView
             android:id="@+id/tvLongitude"
             tools:text="-51.2146267"
+            android:textColor="@color/black"
             android:padding="2dp"
             android:elevation="20dp"
             android:layout_width="wrap_content"
