+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,5 @@
     <color name="black">#000</color>
     <color name="white">#fff</color>
     <color name="transparent">#00FFFFFF</color>
+    <color name="transparentLoader">#B3FFFFFF</color>
 </resources>
